Adding in mailgraph package and updated updaterepo.sh script.

+++ updaterepo.sh
@@ -4,42 +4,51 @@
 #                                                           #
 #############################################################
 
-CENTOS_REPO="https://github.com/DamianZaremba/rpm-centos.git";
-BASE_DIR="/var/www/vhosts/repo.nodehost.co.uk/centos/"
-TMP_DIR=$(mktemp -d /tmp/repoupdate.XXXXX)
+BASE_REPO="git://github.com/DamianZaremba/rpm-%%type%%.git";
+BASE_DIR="/var/www/vhosts/repo.nodehost.co.uk/";
+REPOS=( "centos" "fedora" );
 
-mkdir -p $TMP_DIR
-git clone $CENTOS_REPO $TMP_DIR
-cd $TMP_DIR
-
-echo $TMP_DIR
-for pkg in *;
+for repo in ${REPOS[@]};
 do
-	test -d "$TMP_DIR/$pkg" || continue;
-	cd "$TMP_DIR/$pkg";
-	echo "Starting $pkg"
+	TMP_DIR=$(mktemp -d /tmp/repoupdate.XXXXX);
+	test -d "$BASE_DIR/$repo" || mkdir -p "$BASE_DIR/$repo";
+
+	git clone --quiet --progress $(echo $BASE_REPO | sed "s/%%type%%/$repo/") $TMP_DIR;
+	if [ -d $TMP_DIR ];
+	then
+		cd $TMP_DIR;
+		for pkg in *;
+		do
+			test -d "$TMP_DIR/$pkg" || continue;
+			cd "$TMP_DIR/$pkg";
+			echo "Starting $pkg";
+			for arch in *;
+			do
+				test -d "$TMP_DIR/$pkg/$arch" || continue;
+				cd "$TMP_DIR/$pkg/$arch";
+				echo ".. Starting $arch";
+		
+				test -d "$BASE_DIR/$repo/$arch/" || mkdir -p "$BASE_DIR/$repo/$arch/";
+		
+				echo ".... Copying rpms for $pkg ($arch)";
+				cd "$TMP_DIR/$pkg/$arch";
+				rsync -vr --delete *.rpm "$BASE_DIR/$repo/$arch/";
+			
+				echo "";
+			done
+		done
+	fi
+
+	test -d $TMP_DIR && rm -rf $TMP_DIR;
+	cd "$BASE_DIR/$repo";
+
 	for arch in *;
 	do
-		test -d "$TMP_DIR/$pkg/$arch" || continue;
-		cd "$TMP_DIR/$pkg/$arch";
-		echo ".. Starting $arch";
+		test -d "$BASE_DIR/$repo/$arch" || continue;
+		cd "$BASE_DIR/$repo/$arch";
 
-		test -d "$BASE_DIR/$arch/" || mkdir -p "$BASE_DIR/$arch/";
-
-		echo ".... Copying rpms for $pkg ($arch)";
-		cd "$TMP_DIR/$pkg/$arch";
-		rsync -vr --delete *.rpm "$BASE_DIR/$arch/";
-	
-		echo ""
+		createrepo -d "$BASE_DIR/$repo/$arch";
+		repoview -t "$repo - $arch" "$BASE_DIR/$repo/$arch";
 	done
-done
-rm -rf $TMP_DIR
-cd "$BASE_DIR"
-
-for arch in *;
-do
-	test -d "$BASE_DIR/$arch" || continue;
-	cd "$BASE_DIR/$arch";
-	createrepo "$BASE_DIR/$arch";
 done
 chown -R www-server:www-data "$BASE_DIR";
